% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_types.R
\name{count_distinct_value}
\alias{count_distinct_value}
\title{get a variables' distinct values matrix}
\usage{
count_distinct_value(df, var.list = names(df))
}
\arguments{
\item{df}{the dataframe}

\item{var.list}{the variables to be show,default: all variables}
}
\description{
get a variables' distinct values matrix. this function is typically used to determine 
those discrete variables whose distinct values' number is more than a threshold
}
\examples{
count_distinct_value(p2p)
}

