% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_types.R
\name{define_types}
\alias{define_types}
\title{define the columns' data type}
\usage{
define_types(df, type_fun, vars = NULL)
}
\arguments{
\item{df}{the dataframe}

\item{type_fun}{choose type}

\item{vars}{those variables whose distinct values' numbers is more than a threshold but still need to be discrete type a the maxmimum distinct values' numbers that a discrete variable could have}
}
\description{
This function define the variable's data type by variable's distinct values' numbers
}
\examples{
dis.list<-"auth_key_relation"
define_types(p2p,as.character,3)
}

