% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_types.R
\name{define_types_by_predoc}
\alias{define_types_by_predoc}
\title{define the columns' data type accoding to predefined types doc}
\usage{
define_types_by_predoc(df, pre.types)
}
\arguments{
\item{df}{the dataframe}

\item{pre.types}{predefined types doc}
}
\description{
This function define the columns' data type accoding to predefined doc
}
\examples{

define_types_by_predoc(p2p,pre.types)
}

