% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_handle.R
\name{df_outlier_handle}
\alias{df_outlier_handle}
\title{change those value exit in test dataframe but not in train dataframe}
\usage{
df_outlier_handle(alldata, trainkeys)
}
\arguments{
\item{train}{the traindf}

\item{test}{the testdf}

\item{targ}{the target}

\item{a}{the threshold}
}
\description{
change those value exit in test dataframe but not in train dataframe
}
\examples{
df_outlier_handle(train,test)
}

