% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_strange_sign.R
\name{find_strange_sign}
\alias{find_strange_sign}
\title{find the strange sign in the dataframe}
\usage{
find_strange_sign(df, n)
}
\arguments{
\item{df}{the dataframe need to be exploit}

\item{n}{the numbers of signs to show}
}
\description{
find the strange sign in the dataframe
}
\examples{
find_strange_sign(p2p,100)
}

