% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_handle.R
\name{single_value_count}
\alias{single_value_count}
\title{counting vector's most frequency value's rate}
\usage{
single_value_count(df, targ)
}
\arguments{
\item{df}{the dataframe}

\item{targ}{the label variable}
}
\description{
this function will compute the columns' most frequency value's rate
}
\examples{
single_value_count(p2p,"def")
}

