% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_vector.R
\name{stand_fit}
\alias{stand_fit}
\title{scale vector}
\usage{
stand_fit(df, vars, type = "maxmin")
}
\arguments{
\item{df}{the dataframe}

\item{vars}{the candidates variables}

\item{type}{the scale type}
}
\description{
scale vector
}
\examples{
stand_fit(p2p,con.vars,type='sdm')
}

