% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_vector.R
\name{stand_transform}
\alias{stand_transform}
\title{transform scale vector}
\usage{
stand_transform(df, fit)
}
\arguments{
\item{df}{the dataframe}

\item{fit}{the fitted scaler model}
}
\description{
transform scale vector
}
\examples{
stand_transform(p2p,st.fit)
}

