% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_handle.R
\name{trans_v2v}
\alias{trans_v2v}
\title{define those missing observations}
\usage{
trans_v2v(df, value, change2)
}
\arguments{
\item{df}{the dataframe}

\item{num.fill}{the numeric vector miss filled number}

\item{dis.fill}{the catogary vector miss filled value}
}
\description{
this function will change those filled missing observation into NA format
}
\examples{
trans2miss(p2p,-1,NA)
}

